+++ .vscode/settings.json
@@ -2,7 +2,7 @@
     "python.testing.unittestArgs": [
         "-v",
         "-s",
-        "./*/tests",
+        "./../../tests",
         "-p",
         "test_*.py"
     ],
