+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+### Version 1.1
+
+* Refactoring of the different modules. Added _wta.tennis.models_.
+
+### Version 1
+
+* Initial functionalities
+* Added documentation
