+++ main.py
@@ -1,0 +1,3 @@
+from tennis.wta.parser import ParsePage
+
+ParsePage()
