--- tennis/CHANGELOG.md
@@ -1,4 +1,0 @@
-### Version 1
-
-* Initial functionalities
-* Added documentation
